[Bronze III] Title: 자전거 속도, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2765. 자전거 속도/자전거 속도.cc
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <cmath>
-const double PI = acosl(-1);
+const double PI = std::acos(-1);
 int main(void) {
     for(int i = 1;; i++) {
         double a, c;
